move test transform from simulation to utilities
